requirements.txt
asn1crypto==0.22.0
certifi==2016.2.28
cffi==1.10.0
click==6.7
cryptography==2.0.3
enum34==1.1.6
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
functools32==3.2.3.post2
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.5.3
six==1.11.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
wincertstore==0.2
